Added cron schedules to the release
Each script needs to have its own cron schedule.

The solution found for this problem was to add properties to the release
with the name of the script. These properties will contain the cron
schedule.

Because we can not use variables to read properties in BOSH erb, we need
to find an alternative. The alternative is to create a file (with the
same name as the script), and write the cron schedule into that file.

It's not the most amazing solution ever, but it works for now. Can
always be changed in the future, if something better pops up.